requirements-lambda-simple.txt
boto3>=1.34.0
botocore>=1.34.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
ujson>=5.8.0
structlog>=23.0.0
